pyproject.toml
[project]
name = "healthgptdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "backoff>=2.2.1",
    "google-genai>=1.8.0",
    "google-generativeai>=0.8.4",
    "gradio>=5.23.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jinja2>=3.1.6",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.0",
]
